shopping_list_backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
cfgv==3.4.0
distlib==0.3.8
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==22.0.0
filelock==3.13.1
identify==2.5.33
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
nodeenv==1.8.0
packaging==23.2
phonenumbers==8.13.27
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
psycopg2==2.9.9
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
virtualenv==20.25.0
